+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 **Terraform**
 - Increase minimum supported version to 1.8.
 - Add `refresh`, `lock`, `lockTimeout`, `parallelism`, and `state` params to `plan` method.
+- Add `reconfigure`, `fromModule`, `get`, `lock`, `lockTimeout`, `json`, `lockfile`, and `ignoreRemoteVersion` params to `init` method.
 
 ### 2.2.2
 **General**